+++ gbe/reporting/views/__init__.py
@@ -1,5 +1,7 @@
+from act_techinfo_detail import act_techinfo_detail
 from eval_view import eval_view
 from interest_view import interest_view
 from review_staff_area_view import review_staff_area_view
 from staff_area_view import staff_area_view
 from welcome_letter_view import WelcomeLetterView
+from review_act_techinfo import review_act_techinfo
