--- gbe/templates/gbe/profile_view.tmpl
@@ -1,11 +1,0 @@
-{% extends 'base.tmpl' %}
-
-{% block title %}
-   Performer Edit
-{% endblock %}
-
-{% block content %}
-
-   <p class="vanilla">View a Profile </p>
-   {{ form }}
-{% endblock %}
